faet(MainPage): tsx 확장자로변경
